+++ controller/general_albaranes_prov.php
@@ -21,6 +21,8 @@ require_once 'model/albaran_proveedor.php';
 
 class general_albaranes_prov extends fs_controller
 {
+   public $buscar_lineas;
+   public $lineas;
    public $resultados;
    public $offset;
    
@@ -31,41 +33,49 @@ class general_albaranes_prov extends fs_controller
    
    protected function process()
    {
-      $albaran = new albaran_proveedor();
-      $this->custom_search = TRUE;
-      
-      $npage = $this->page->get('general_nuevo_albaran');
-      if($npage)
-         $this->buttons[] = new fs_button('b_nuevo_albaran', 'nuevo', $npage->url());
-      
-      if( isset($_GET['delete']) )
+      if( isset($_POST['buscar_lineas']) )
+         $this->buscar_lineas();
+      else
       {
-         $alb1 = new albaran_proveedor();
-         $alb1 = $alb1->get($_GET['delete']);
-         if($alb1)
+         $albaran = new albaran_proveedor();
+         $this->custom_search = TRUE;
+         
+         $npage = $this->page->get('general_nuevo_albaran');
+         if($npage)
+            $this->buttons[] = new fs_button('b_nuevo_albaran', 'nuevo', $npage->url());
+         
+         $this->buttons[] = new fs_button('b_buscar_lineas', 'lineas', '#', '', 'img/zoom.png');
+         
+         if( isset($_GET['delete']) )
          {
-            if( $alb1->delete() )
-               $this->new_message("Albarán ".$alb1->codigo." borrado correctamente.");
+            $alb1 = new albaran_proveedor();
+            $alb1 = $alb1->get($_GET['delete']);
+            if($alb1)
+            {
+               if( $alb1->delete() )
+                  $this->new_message("Albarán ".$alb1->codigo." borrado correctamente.");
+               else
+                  $this->new_error_msg("¡Imposible borrar el albarán!");
+            }
             else
-               $this->new_error_msg("¡Imposible borrar el albarán!");
+               $this->new_error_msg("¡Albarán no encontrado!");
          }
+         
+         if( isset($_GET['offset']) )
+            $this->offset = intval($_GET['offset']);
          else
-            $this->new_error_msg("¡Albarán no encontrado!");
+            $this->offset = 0;
+         
+         if($this->query != '')
+            $this->resultados = $albaran->search($this->query, $this->offset);
+         else
+            $this->resultados = $albaran->all($this->offset);
       }
-      
-      if( isset($_GET['offset']) )
-         $this->offset = intval($_GET['offset']);
-      else
-         $this->offset = 0;
-      
-      if($this->query != '')
-         $this->resultados = $albaran->search($this->query, $this->offset);
-      else
-         $this->resultados = $albaran->all($this->offset);
    }
    
-   public function version() {
-      return parent::version().'-3';
+   public function version()
+   {
+      return parent::version().'-4';
    }
    
    public function anterior_url()
@@ -86,6 +96,16 @@ class general_albaranes_prov extends fs_controller
       else if($this->query=='' AND count($this->resultados)==FS_ITEM_LIMIT)
          $url = $this->url()."&offset=".($this->offset+FS_ITEM_LIMIT);
       return $url;
+   }
+   
+   public function buscar_lineas()
+   {
+      /// cambiamos la plantilla HTML
+      $this->template = 'ajax/general_lineas_albaranes_prov';
+      
+      $this->buscar_lineas = $_POST['buscar_lineas'];
+      $linea = new linea_albaran_proveedor();
+      $this->lineas = $linea->search($this->buscar_lineas);
    }
 }
